+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.5.8] - 2020-10-16
+- Add product attribute for product tax codes
+- Fix mismatched tax codes for configurable products
+- Fix error that prevented orders with deleted products from syncing to TaxJar
+
 ## [1.5.7] - 2020-09-03
 - Improve the display of the admin configuration page
 
@@ -206,7 +211,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Initial release of our Magento 2 extension.** Sales tax calculations at checkout with backup zip-based rates powered by TaxJar. Supports product exemptions, shipping taxability, sourcing logic, and international calculations in more than 30 countries.
 - **Special promo sales tax calculations for Magento merchants.** Existing M2 beta users must upgrade to this version to receive special promo calculations at checkout using our new API endpoint.
 
-[Unreleased]: https://github.com/taxjar/taxjar-magento2-extension/compare/v1.5.7...HEAD
+[Unreleased]: https://github.com/taxjar/taxjar-magento2-extension/compare/v1.5.8...HEAD
+[1.5.8]: https://github.com/taxjar/taxjar-magento2-extension/compare/v1.5.7...v1.5.8
 [1.5.7]: https://github.com/taxjar/taxjar-magento2-extension/compare/v1.5.6...v1.5.7
 [1.5.6]: https://github.com/taxjar/taxjar-magento2-extension/compare/v1.5.5...v1.5.6
 [1.5.5]: https://github.com/taxjar/taxjar-magento2-extension/compare/v1.5.4...v1.5.5